requirements.txt
boto3==1.16.12
botocore==1.19.12
jmespath==0.10.0
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
